+++ packages/codegen/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Account for storage entries that use opaque hashers.
+
 ## 0.12.8 - 2024-11-15
 
 ### Fixed